+++ crates/contract/Cargo.toml
@@ -98,6 +98,7 @@ near-mpc-contract-interface = { workspace = true, features = [
     "k256",
     "near",
 ] }
+near-mpc-sdk = { workspace = true }
 near-mpc-signature-verifier = { workspace = true }
 near-sdk = { workspace = true }
 rand = { workspace = true, optional = true }
