+++ crates/near-mpc-contract-interface/src/types/foreign_chain.rs
@@ -1955,27 +1955,6 @@ pub struct ProviderConfig {
     pub chain_routing: ChainRouting,
 }
 
-/// Reasons a [`ChainEntry`] proposal fails validation. [`NonEmptyBTreeMap`]
-/// already enforces non-empty + unique-[`ProviderId`] at deserialize time,
-/// so those cases are absent here.
-#[derive(Debug, Clone, Eq, PartialEq, thiserror::Error)]
-pub enum ChainEntryValidationError {
-    #[error("ChainEntry.quorum must be >= 1")]
-    ZeroQuorum,
-    #[error(
-        "ChainEntry.quorum ({quorum}) exceeds providers.len() ({providers_len}) — RPC response quorum is unreachable"
-    )]
-    QuorumExceedsProviders { quorum: u64, providers_len: u64 },
-    #[error(
-        "ChainRouting::PathSegment.segment for provider_id {provider_id:?} must not contain '/'"
-    )]
-    PathSegmentContainsSlash { provider_id: String },
-    #[error(
-        "ChainRouting::QueryParam.name collides with AuthScheme::Query.name {name:?} for provider_id {provider_id:?}"
-    )]
-    QueryParamCollidesWithAuth { provider_id: String, name: String },
-}
-
 /// Stored state for one chain in the on-chain whitelist: a non-empty map from
 /// [`ProviderId`] to that provider's per-chain configuration, plus the RPC response
 /// quorum nodes should use when querying. Returned by the
@@ -1996,52 +1975,10 @@ pub struct ChainEntry {
     pub quorum: u64,
 }
 
-impl ChainEntry {
-    /// Checks the rules the contract enforces before a proposal can apply; see [`ChainEntryValidationError`].
-    pub fn validate(&self) -> Result<(), ChainEntryValidationError> {
-        let ChainEntry { providers, quorum } = self;
-        if *quorum == 0 {
-            return Err(ChainEntryValidationError::ZeroQuorum);
-        }
-        let providers_len = u64::try_from(providers.len())
-            .expect("usize is at most 64 bits on all supported targets");
-        if *quorum > providers_len {
-            return Err(ChainEntryValidationError::QuorumExceedsProviders {
-                quorum: *quorum,
-                providers_len,
-            });
-        }
-        for (id, config) in providers.iter() {
-            if let ChainRouting::PathSegment { segment } = &config.chain_routing
-                && segment.contains('/')
-            {
-                return Err(ChainEntryValidationError::PathSegmentContainsSlash {
-                    provider_id: id.0.clone(),
-                });
-            }
-            if let (
-                ChainRouting::QueryParam {
-                    name: routing_name, ..
-                },
-                AuthScheme::Query { name: auth_name },
-            ) = (&config.chain_routing, &config.auth_scheme)
-                && routing_name == auth_name
-            {
-                return Err(ChainEntryValidationError::QueryParamCollidesWithAuth {
-                    provider_id: id.0.clone(),
-                    name: auth_name.clone(),
-                });
-            }
-        }
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 #[expect(non_snake_case)]
 mod tests {
     use super::*;
-    use assert_matches::assert_matches;
     use rstest::rstest;
 
     #[test]
@@ -2614,124 +2551,5 @@ mod tests {
 
         // Then
         assert_ne!(hash_7_bits, hash_8_bits);
-    }
-
-    fn provider(id: &str) -> (ProviderId, ProviderConfig) {
-        (
-            ProviderId(id.to_string()),
-            ProviderConfig {
-                base_url: format!("https://{id}.example.com"),
-                auth_scheme: AuthScheme::None,
-                chain_routing: ChainRouting::Embedded,
-            },
-        )
-    }
-
-    /// Build a [`ChainEntry`] from a list of provider id stubs and a quorum.
-    fn chain_entry(ids: &[&str], quorum: u64) -> ChainEntry {
-        let providers: BTreeMap<ProviderId, ProviderConfig> =
-            ids.iter().map(|id| provider(id)).collect();
-        ChainEntry {
-            providers: NonEmptyBTreeMap::try_from(providers)
-                .expect("test setup: providers must be non-empty"),
-            quorum,
-        }
-    }
-
-    #[test]
-    fn validate_chain_entry__should_reject_zero_quorum() {
-        // Given
-        let entry = chain_entry(&["alchemy"], 0);
-
-        // When
-        let err = entry.validate().unwrap_err();
-
-        // Then
-        assert_matches!(err, ChainEntryValidationError::ZeroQuorum);
-    }
-
-    #[test]
-    fn validate_chain_entry__should_reject_quorum_exceeding_providers_count() {
-        // Given
-        let entry = chain_entry(&["alchemy"], 2);
-
-        // When
-        let err = entry.validate().unwrap_err();
-
-        // Then
-        assert_matches!(
-            err,
-            ChainEntryValidationError::QuorumExceedsProviders {
-                quorum: 2,
-                providers_len: 1,
-            }
-        );
-    }
-
-    #[test]
-    fn validate_chain_entry__should_reject_path_segment_containing_slash() {
-        // Given
-        let entry = ChainEntry {
-            providers: NonEmptyBTreeMap::new(
-                ProviderId("ankr".to_string()),
-                ProviderConfig {
-                    base_url: "https://rpc.ankr.com".to_string(),
-                    auth_scheme: AuthScheme::None,
-                    chain_routing: ChainRouting::PathSegment {
-                        segment: "eth/sepolia".to_string(),
-                    },
-                },
-            ),
-            quorum: 1,
-        };
-
-        // When
-        let err = entry.validate().unwrap_err();
-
-        // Then
-        assert_matches!(
-            err,
-            ChainEntryValidationError::PathSegmentContainsSlash { provider_id } if provider_id == "ankr"
-        );
-    }
-
-    #[test]
-    fn validate_chain_entry__should_reject_query_param_colliding_with_auth_query() {
-        // Given
-        let entry = ChainEntry {
-            providers: NonEmptyBTreeMap::new(
-                ProviderId("drpc".to_string()),
-                ProviderConfig {
-                    base_url: "https://lb.drpc.org/ogrpc".to_string(),
-                    auth_scheme: AuthScheme::Query {
-                        name: "key".to_string(),
-                    },
-                    chain_routing: ChainRouting::QueryParam {
-                        name: "key".to_string(),
-                        value: "ethereum".to_string(),
-                    },
-                },
-            ),
-            quorum: 1,
-        };
-
-        // When
-        let err = entry.validate().unwrap_err();
-
-        // Then
-        assert_matches!(
-            err,
-            ChainEntryValidationError::QueryParamCollidesWithAuth { provider_id, name }
-                if provider_id == "drpc" && name == "key"
-        );
-    }
-
-    #[test]
-    fn validate_chain_entry__should_accept_well_formed_entry() {
-        // Given
-        let entry = chain_entry(&["alchemy", "ankr"], 2);
-
-        // When / Then
-        entry.validate().expect("well-formed entry should validate");
     }
 }
