+++ crates/near-mpc-sdk/Cargo.toml
@@ -20,6 +20,7 @@ derive_more = { workspace = true }
 near-mpc-bounded-collections = { workspace = true }
 near-mpc-contract-interface = { workspace = true }
 near-mpc-signature-verifier = { workspace = true }
+thiserror = { workspace = true }
 
 [dev-dependencies]
 assert_matches = { workspace = true }
