+++ crates/proposal-generator/Cargo.toml
@@ -13,6 +13,7 @@ clap.workspace = true
 hex.workspace = true
 near-mpc-bounded-collections.workspace = true
 near-mpc-contract-interface.workspace = true
+near-mpc-sdk.workspace = true
 serde.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
